feat: Refactor getApiKey from landing package to shared package.
