+++ packages/landing/src/__test__/config.test.ts
@@ -1,0 +1,9 @@
+import o from 'ospec';
+import { Config } from '../config';
+
+o.spec('Config', () => {
+    o('should return the same api key', () => {
+        const keyA = Config.ApiKey;
+        o(keyA).equals(Config.ApiKey);
+    });
+});
